[Rename] GetFriendCommand -> GetFriendResponse
