+++ BE/src/main/java/com/d201/fundingift/friend/domain/port/FriendExternalPort.java
@@ -1,10 +1,10 @@
 package com.d201.fundingift.friend.domain.port;
 
-import com.d201.fundingift.friend.dto.GetFriendCommand;
+import com.d201.fundingift.friend.dto.response.GetFriendResponse;
 
 import java.util.List;
 
 public interface FriendExternalPort {
 
-    List<GetFriendCommand> getFriends(Long consumerId); // 친구 목록 불러오기
+    List<GetFriendResponse> getFriends(Long consumerId); // 친구 목록 불러오기
 }
